+++ src/physics/kinematics.h
@@ -2,16 +2,16 @@
 
 #include <math.h>
 
-float phyc_position(float si, float v, float a, float t);
-float phyc_velocity(float vi, float a, float t);
-float phyc_medium_velocity(float pi, float pf, float ti, float tf);
-float phyc_medium_acceleration(float vi, float vf, float t);
-float phyc_magnitude(float bx, float by);
-float phyc_displacement_x_y(float ai, float af, float bi, float bf);
-float phyc_magnitude_velocity(float vx, float vy);
-float phyc_magnitude_acceleration(float ax, float ay);
-float phyc_decompose_x(float b, float angle);
-float phyc_decompose_y(float b, float angle);
-float phyc_angle(float fx, float fy);
-float phyc_degree_to_radian(float degree);
-float phyc_radian_to_degree(float radian);
+double phyc_position(double si, double v, double a, double t);
+double phyc_velocity(double vi, double a, double t);
+double phyc_medium_velocity(double pi, double pf, double ti, double tf);
+double phyc_medium_acceleration(double vi, double vf, double t);
+double phyc_magnitude(double bx, double by);
+double phyc_displacement_x_y(double ai, double af, double bi, double bf);
+double phyc_magnitude_velocity(double vx, double vy);
+double phyc_magnitude_acceleration(double ax, double ay);
+double phyc_decompose_x(double b, double angle);
+double phyc_decompose_y(double b, double angle);
+double phyc_angle(double fx, double fy);
+double phyc_degree_to_radian(double degree);
+double phyc_radian_to_degree(double radian);
